requirements.txt
litestar==2.11.0
openai==1.54.4
google-genai==1.43.0
anthropic==0.39.0
httpx>=0.28.1,<1.0.0
python-multipart==0.0.9
python-dotenv==1.0.1
uvicorn[standard]==0.30.6
pillow==10.4.0
requests==2.32.3
beautifulsoup4==4.12.3
sqlalchemy[asyncio]>=2.0.0
aiosql>=10.0
aiosqlite>=0.19.0
asyncpg>=0.29.0
supabase>=2.0.0
